Refactoring: renaming, package moving, encapsulation
