+++ src/main/java/com/leokom/chess/players/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Provide players for the Chess game.
+ * The goal of this package is to decide who will play chess in the current game
+ */
+package com.leokom.chess.players;
